fixtures: switch LP64-only assumptions to long long for LLP64 portability
CI on the windows-latest and windows-11-arm runners surfaced seven
fixtures that assume `sizeof(long) == 8`. They were green when c5
treated `long` as 8 bytes everywhere; once `068fc69` made `long` 4
bytes on Windows targets they fail with sizeof mismatches and
ACCESS_VIOLATIONs (the latter from struct nodes that pack a pointer
into a `long` cell -- the truncated address dereferences into garbage).

The portable fix is to use `long long` (8 bytes on every C99 target)
when the slot needs to hold an 8-byte value or a pointer.

* `linked_list.c`, `binary_search_tree.c`, `bst_free.c`: all `long *`
  node arrays carry pointer cells, so storage is now `long long *`
  with `long long` value cells. Every `(long)` / `(long *)` cast
  becomes `(long long)` / `(long long *)`.
* `float_pointer_basics.c`: 64-bit double bit patterns are now
  written through `long long *` rather than `long *` to keep the
  high half on Windows.
* `integer_ops_exhaustive.c`: 64-bit width band moved from
  `(unsigned) long` to `(unsigned) long long`. Comment header
  updated to reflect the four bands (char, short, int, long long).
* `integer_suffixes.c`: `big_dec` / `big_hex` are 12-digit decimals
  / 33-bit hex, so their storage is now `long long`. The `LL`
  suffix on the literals already matched.
* `sizeof_basic.c`, `sizeof_expr.c`: `sizeof(long)` and `sizeof(*long_ptr)`
  checks now branch on `__BADC_WINDOWS__` (4 vs 8). Pointer sizes
  stay 8 unconditionally.

Also retires `deferred_long_width.c` and the `long_width_per_target`
deferred test -- their gap closed when `068fc69` landed Ty::LongLong
and target-aware sizing. The new `fixtures/c/long_long_distinct.c`
(in 3417c04) covers the same per-target shape and runs in all five
backend parity tables.

355 tests pass on macOS arm64 (was 355 / 10 ignored, now 355 / 9
ignored after retiring the per-target-long-width deferred test).
sqlite3 smoke green at -O and noO.

+++ fixtures/c/binary_search_tree.c
@@ -1,36 +1,38 @@
 #include <stdlib.h>
 
-// Tree nodes are packed as long[3]: [value, left, right]. The cells
-// hold pointer-shaped values, so under  (where `int` is 4 bytes
-// and a pointer is 8) the storage type has to be `long`. We cast
-// between `long *` and `long` at every store/load to mix value and
-// pointer cells in the same array; the c4 dialect has no long**
-// so flat-array nodes are the natural shape.
-long* insert(long *root, long val) {
+// Tree nodes are packed as long long[3]: [value, left, right]. The
+// cells hold pointer-shaped values, so they need to be 8 bytes
+// everywhere (int is 4 bytes, and `long` is only 4 bytes on Windows
+// LLP64). `long long` is 8 bytes on every target. We cast between
+// `long long *` and `long long` at every store/load to mix value and
+// pointer cells in the same array; the c4 dialect has no
+// long-long-pointer-pointers so flat-array nodes are the natural
+// shape.
+long long* insert(long long *root, long long val) {
     if (root == 0) {
-        root = malloc(sizeof(long) + 2 * sizeof(long *));
+        root = malloc(sizeof(long long) + 2 * sizeof(long long *));
         root[0] = val;
         root[1] = 0;
         root[2] = 0;
         return root;
     }
     if (val < root[0]) {
-        root[1] = (long)insert((long *)root[1], val);
+        root[1] = (long long)insert((long long *)root[1], val);
     } else {
-        root[2] = (long)insert((long *)root[2], val);
+        root[2] = (long long)insert((long long *)root[2], val);
     }
     return root;
 }
 
-int search(long *root, long val) {
+int search(long long *root, long long val) {
     if (root == 0) return 0;
     if (root[0] == val) return 1;
-    if (val < root[0]) return search((long *)root[1], val);
-    return search((long *)root[2], val);
+    if (val < root[0]) return search((long long *)root[1], val);
+    return search((long long *)root[2], val);
 }
 
 int main() {
-    long *root;
+    long long *root;
     root = 0;
     root = insert(root, 50);
     insert(root, 30);
