+++ fixtures/c/float_pointer_basics.c
@@ -7,8 +7,9 @@
 // FP value in an 8-byte slot; the IEEE 754 32-bit narrowing for
 // `float` is future work).
 //
-// 32-bit bit patterns round-trip through `int` (which is 4 bytes
-//) and 64-bit bit patterns through `long`.
+// 32-bit bit patterns round-trip through `int` (4 bytes everywhere)
+// and 64-bit bit patterns through `long long` (8 bytes on every
+// target, unlike `long` which is 4 bytes on Windows LLP64).
 int main() {
     float *fa;
     double *da;
@@ -32,15 +33,16 @@ int main() {
     // store). The single-precision pattern fits in `int`.
     *(int *)&fa[0] = 0x3f800000;        // bit pattern of 1.0f
     *(int *)&fa[1] = 0x40000000;        // bit pattern of 2.0f
-    // Double's 64-bit pattern needs a real `long` slot
-    // -- `int` is 4 bytes and would truncate the high half.
-    *(long *)&da[0] = 0x3ff0000000000000; // bit pattern of 1.0
-    *(long *)&da[1] = 0x4000000000000000; // bit pattern of 2.0
+    // Double's 64-bit pattern needs a real 8-byte slot -- `int` is
+    // 4 bytes and would truncate the high half, and so is `long`
+    // on Windows LLP64. `long long` is 8 bytes on every target.
+    *(long long *)&da[0] = 0x3ff0000000000000; // bit pattern of 1.0
+    *(long long *)&da[1] = 0x4000000000000000; // bit pattern of 2.0
 
     if (*(int *)&fa[0] != 0x3f800000) return 3;
     if (*(int *)&fa[1] != 0x40000000) return 4;
-    if (*(long *)&da[0] != 0x3ff0000000000000) return 5;
-    if (*(long *)&da[1] != 0x4000000000000000) return 6;
+    if (*(long long *)&da[0] != 0x3ff0000000000000) return 5;
+    if (*(long long *)&da[1] != 0x4000000000000000) return 6;
 
     free((char *)fa);
     free((char *)da);
